✨ feat: widgets/editor 의존을 줄이도록 도메인 타입과 유틸을 분리
- editor 도메인 타입을 entities/editor/model로 이동

- editor 상태 유틸을 entities/editor/model로 이동

- entities 계층의 widgets/editor 타입 의존 제거

- widgets/editor는 호환용 re-export로 정리
