+++ src/app/api/(files)/images/route.ts
@@ -1,9 +1,9 @@
 import { EDITOR_API_ERROR_MESSAGE } from '@/entities/editor/model/editor-api-error';
+import type { EditorContentType } from '@/entities/editor/model/editor-types';
 import { uploadImageFile } from '@/features/upload-image-file/api/upload-image-file';
 import { API_INTERNAL_ERROR_MESSAGE } from '@/shared/lib/http/api-error-catalog';
 import { createApiErrorResponse } from '@/shared/lib/http/api-response';
 import { runJsonRoute } from '@/shared/lib/http/run-json-route';
-import type { EditorContentType } from '@/widgets/editor/model/editor-core.types';
 
 const contentTypes: EditorContentType[] = ['article', 'project', 'resume'];
