+++ src/widgets/editor/index.ts
@@ -1,13 +1,12 @@
+export type { EditorCoreProps, PublishPanelProps } from './model/editor-core.types';
+export { EditorCore } from './ui/editor-core';
+export { PublishPanel } from './ui/publish-panel';
 export type {
   DraftSaveResult,
   EditorContentType,
-  EditorCoreProps,
   EditorPublicationState,
   EditorState,
   Locale,
-  PublishPanelProps,
   PublishSettings,
   TranslationField,
-} from './model/editor-core.types';
-export { EditorCore } from './ui/editor-core';
-export { PublishPanel } from './ui/publish-panel';
+} from '@/entities/editor/model/editor-types';
